refactor: simplify the whole construct
