+++ lib/src/popover/popover.dart
@@ -1,3 +1,2 @@
-export 'sbb_anchored_overlay_builder.dart';
 export 'sbb_popover.dart';
 export 'sbb_popover_direction.dart';
